+++ analytics.js
@@ -898,15 +898,15 @@ exports.parse = function(url){
   a.href = url;
   return {
     href: a.href,
-    host: a.host,
-    port: a.port,
+    host: a.host || location.host,
+    port: ('0' === a.port || '' === a.port) ? port(a.protocol) : a.port,
     hash: a.hash,
-    hostname: a.hostname,
-    pathname: a.pathname,
-    protocol: a.protocol,
+    hostname: a.hostname || location.hostname,
+    pathname: a.pathname.charAt(0) != '/' ? '/' + a.pathname : a.pathname,
+    protocol: !a.protocol || ':' == a.protocol ? location.protocol : a.protocol,
     search: a.search,
     query: a.search.slice(1)
-  }
+  };
 };
 
 /**
@@ -918,9 +918,7 @@ exports.parse = function(url){
  */
 
 exports.isAbsolute = function(url){
-  if (0 == url.indexOf('//')) return true;
-  if (~url.indexOf('://')) return true;
-  return false;
+  return 0 == url.indexOf('//') || !!~url.indexOf('://');
 };
 
 /**
@@ -932,7 +930,7 @@ exports.isAbsolute = function(url){
  */
 
 exports.isRelative = function(url){
-  return ! exports.isAbsolute(url);
+  return !exports.isAbsolute(url);
 };
 
 /**
@@ -945,10 +943,29 @@ exports.isRelative = function(url){
 
 exports.isCrossDomain = function(url){
   url = exports.parse(url);
-  return url.hostname != location.hostname
-    || url.port != location.port
-    || url.protocol != location.protocol;
+  return url.hostname !== location.hostname
+    || url.port !== location.port
+    || url.protocol !== location.protocol;
 };
+
+/**
+ * Return default port for `protocol`.
+ *
+ * @param  {String} protocol
+ * @return {String}
+ * @api private
+ */
+function port (protocol){
+  switch (protocol) {
+    case 'http:':
+      return 80;
+    case 'https:':
+      return 443;
+    default:
+      return location.port;
+  }
+}
+
 });
 require.register("component-bind/index.js", function(exports, require, module){
 
@@ -3638,9 +3655,15 @@ Evergage.prototype.load = function (callback) {
  */
 
 Evergage.prototype.page = function (category, name, properties, options) {
+  if (name) push('namePage', name);
+
+  properties = properties || {};
+  each(properties, function(key, value) {
+    push('setCustomField', key, value, 'page');
+  });
+
   window.Evergage.init(true);
 };
-
 
 /**
  * Trait aliases.
@@ -8513,11 +8536,19 @@ exports.is = function (string, strict) {
 });
 require.register("segmentio-isodate-traverse/index.js", function(exports, require, module){
 
-var clone = require('clone')
-  , each = require('each')
-  , is = require('is')
-  , isodate = require('isodate');
+var is = require('is');
+var isodate = require('isodate');
 
+var clone;
+var each;
+
+try {
+  clone = require('clone');
+  each = require('each');
+} catch (err) {
+  clone = require('clone-component');
+  each = require('each-component');
+}
 
 /**
  * Expose `traverse`.
@@ -9513,7 +9544,7 @@ var analytics = module.exports = exports = new Analytics();
  * Expose `VERSION`.
  */
 
-exports.VERSION = '1.0.9';
+exports.VERSION = '1.1.0';
 
 
 /**
@@ -9869,13 +9900,18 @@ Analytics.prototype.page = function (category, name, properties, options, fn) {
   if (is.object(category)) options = properties, properties = name, name = category = null;
   if (is.string(category) && !is.string(name)) name = category, category = null;
 
-  properties = clone(properties) || {};
-  defaults(properties, {
+  var defs = {
     path: canonicalPath(),
     referrer: document.referrer,
     title: document.title,
     url: location.href
-  });
+  };
+
+  if (name) defs.name = name;
+  if (category) defs.category = category;
+
+  properties = clone(properties) || {};
+  defaults(properties, defs);
 
   this._invoke('page', category, name, properties, options);
   this._callback(fn);